Add essential elements to index + Update README
This commit adds sections 2-4 as well as the footer. Some of the
elements have some class names attached to them. This will be expanded
on when working on styling the html file. In addition I added an image
to the README so that it will show what the intended end result should
look like.