Link CSS + Style header
This commit adds a stylesheet to the html file and adds some classes
so that it is easier to find in the CSS file. On the CSS hand, there
is styling for the body and elements in the header. From this commit,
the landing page should have a correctly positioned header as well as
some elements having the proper format